moved validate_spatialdata into utils and removed io module entirely
