--- src/segtraq/io/__init__.py
@@ -1,5 +1,0 @@
-from .io import validate_spatialdata
-
-__all__ = [
-    "validate_spatialdata",
-]
