--- src/segtraq/segtraq_metrics/__init__.py
@@ -1,5 +1,0 @@
-from .SegtraqMetrics import SegTraqer
-
-__all__ = [
-    "SegTraqer",
-]
